Clean run.sh and add check for no-file in backup.sh
This commit also has work on circleci config as well
as various other minor changes in the backup script.

+++ dockerfiles/backup.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env sh
+#!/usr/bin/env bash
 # Extremely dangerous, hardcoded and barebone backup script
 
 set -eo pipefail
@@ -11,21 +11,28 @@ backup() {
 
 # Restore from host
 restore() {
+    if [ ! -f "$1" ]; 
+    then
+        echo "No arg given or file does not exist"
+        exit 1
+    fi
+
     echo "Restoring data..."
     docker exec -i minitwit-db psql -U embu -d minitwit-db -c \
         "DROP SCHEMA PUBLIC CASCADE; CREATE SCHEMA PUBLIC;" &&
-    docker exec -i minitwit-db psql -U embu -d minitwit-db < backup.sql
+    docker exec -i minitwit-db psql -U embu -d minitwit-db < $1
 }
 
 case $1 in
     backup)
-        backup
-        ;;
+        backup $2 ;;
     restore)
-        restore
-        ;;
+        restore $2 ;;
     *)
-        echo "Command not found." >&2
+        echo -e "Usage:\n"
+        echo "arg1      arg2        action"
+        echo "backup                will backup minitwit-db to app root as backup.sql"
+        echo "restore   <path>      will delete minitwit-db in docker and restore using <path>"
         exit 1
         ;;
 esac
